chore: Proposal for new module based file structure (#13396)
