+++ apps/web/modules/bookings/lib/validStatuses.ts
@@ -1,0 +1,1 @@
+export const validStatuses = ["upcoming", "recurring", "past", "cancelled", "unconfirmed"] as const;
