+++ apps/web/pages/booking/[uid]/index.tsx
@@ -1,0 +1,2 @@
+export { default } from "~/bookings/views/bookings-single-view";
+export { getServerSideProps } from "~/bookings/views/bookings-single-view.getServerSideProps";
